+++ src/Settings.ts
@@ -7,6 +7,7 @@ import * as helpers from "./helpers";
 import * as log from "./Log";
 
 import IMqttManagerConfigJson from "./handlers/mqttManager/IMqttManagerConfigJson";
+import IPushbulletManagerConfigJson from "./handlers/pushbulletManager/IPushbulletManagerConfigJson";
 import IPushoverManagerConfigJson from "./handlers/pushoverManager/IPushoverManagerConfigJson";
 import ISettingsConfigJson from "./types/ISettingsConfigJson";
 import ITelegramManagerConfigJson from "./handlers/telegramManager/ITelegramManagerConfigJson";
@@ -20,6 +21,7 @@ export let port: number;
 export let processExistingImages: boolean;
 export let purgeAge: number;
 export let purgeInterval: number;
+export let pushbullet: IPushbulletManagerConfigJson;
 export let pushover: IPushoverManagerConfigJson;
 export let telegram: ITelegramManagerConfigJson;
 export let verbose: boolean;
@@ -62,6 +64,7 @@ export function loadConfiguration(configFilePaths: string[]): string {
   processExistingImages = settingsConfigJson.processExistingImages ?? false;
   purgeAge = settingsConfigJson.purgeAge ?? 30;
   purgeInterval = settingsConfigJson.purgeInterval ?? 60;
+  pushbullet = settingsConfigJson.pushbullet;
   pushover = settingsConfigJson.pushover;
   telegram = settingsConfigJson.telegram;
   verbose = settingsConfigJson.verbose ?? false;