+++ src/TriggerManager.ts
@@ -7,6 +7,7 @@ import * as log from "./Log";
 
 import ITriggerConfigJson from "./types/ITriggerConfigJson";
 import MqttHandlerConfig from "./handlers/mqttManager/MqttHandlerConfig";
+import PushbulletConfig from "./handlers/pushbulletManager/PushbulletConfig";
 import PushoverConfig from "./handlers/pushoverManager/PushoverConfig";
 import Rect from "./Rect";
 import TelegramConfig from "./handlers/telegramManager/TelegramConfig";
@@ -86,17 +87,20 @@ export function loadConfiguration(configFilePaths: string[]): string {
     );
 
     // Set up the handlers
-    if (triggerJson.handlers.webRequest) {
-      configuredTrigger.webRequestHandlerConfig = new WebRequestConfig(triggerJson.handlers.webRequest);
-    }
     if (triggerJson.handlers.mqtt) {
       configuredTrigger.mqttHandlerConfig = new MqttHandlerConfig(triggerJson.handlers.mqtt);
+    }
+    if (triggerJson.handlers.pushbullet) {
+      configuredTrigger.pushbulletConfig = new PushbulletConfig(triggerJson.handlers.pushbullet);
+    }
+    if (triggerJson.handlers.pushover) {
+      configuredTrigger.pushoverConfig = new PushoverConfig(triggerJson.handlers.pushover);
     }
     if (triggerJson.handlers.telegram) {
       configuredTrigger.telegramConfig = new TelegramConfig(triggerJson.handlers.telegram);
     }
-    if (triggerJson.handlers.pushover) {
-      configuredTrigger.pushoverConfig = new PushoverConfig(triggerJson.handlers.pushover);
+    if (triggerJson.handlers.webRequest) {
+      configuredTrigger.webRequestHandlerConfig = new WebRequestConfig(triggerJson.handlers.webRequest);
     }
 
     return configuredTrigger;