+++ src/handlers/pushbulletManager/IPushoverConfigJson.ts
@@ -1,0 +1,11 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (c) Neil Enns. All rights reserved.
+ *  Licensed under the MIT License. See LICENSE in the project root for license information.
+ *--------------------------------------------------------------------------------------------*/
+export default interface IPushbulletConfigJson {
+  annotateImage?: boolean;
+  caption?: string;
+  cooldownTime?: number;
+  enabled?: boolean;
+  title?: string;
+}